+++ hutool-log/src/main/java/com/xiaoleilu/hutool/log/dialect/tinylog/TinyLogFactory.java
@@ -1,0 +1,32 @@
+package com.xiaoleilu.hutool.log.dialect.tinylog;
+
+import com.xiaoleilu.hutool.log.Log;
+import com.xiaoleilu.hutool.log.LogFactory;
+
+/**
+ * <a href="http://www.tinylog.org/">TinyLog</a> log.<br>
+ * 
+ * @author Looly
+ *
+ */
+public class TinyLogFactory extends LogFactory {
+	
+	/**
+	 * 构造
+	 */
+	public TinyLogFactory() {
+		super("TinyLog");
+		checkLogExist(org.pmw.tinylog.Logger.class);
+	}
+
+	@Override
+	public Log getLog(String name) {
+		return new TinyLog(name);
+	}
+
+	@Override
+	public Log getLog(Class<?> clazz) {
+		return new TinyLog(clazz);
+	}
+
+}